Check before unlink memofile in destroy for nomemo mode.

+++ rubylib/maildeliv-mailfilter.rb
@@ -51,7 +51,9 @@ module MailFilter
   
   def destroymail
     STDERR.puts "destroymail() : Destroy a mail from #{@mailobj.address}"
-    File.unlink @memofile
+    if [[ File.exist?(@memofile) ]]
+      File.unlink @memofile
+    end
     throw :filtering
   end
